requirements.txt
PySide6==6.3
python-rapidjson>=1.6
